Feature/su 104 user endpoints

+++ backend/src/schoolAdmin/schoolAdmin.service.ts
@@ -546,6 +546,26 @@ export class SchoolAdminService {
           },
         },
       });
+
+      await prisma.personRoles.create({
+        data: {
+          persons: {
+            connect: {
+              personId: personId,
+            },
+          },
+          roles: {
+            connect: {
+              roleId: 2,
+            },
+          },
+          schools: {
+            connect: {
+              schoolId: schoolId,
+            },
+          },
+        },
+      });
     } catch (err) {
       if (err.code === 'P2002') {
         return RETURN_DATA.ALREADY_EXISTS;
@@ -584,6 +604,24 @@ export class SchoolAdminService {
     const schoolId = await this.databaseService.getSchoolIdByUUID(schoolUUID);
     const personId = await this.databaseService.getPersonIdByUUID(personUUID);
 
+    const role = await this.databaseService.getUserRoles(personId);
+    console.log(role);
+
+    const isOnlyUser = await prisma.schoolPersons.findFirst({
+      where: {
+        persons: {
+          personId: personId,
+        },
+        schools: {
+          schoolId: schoolId,
+        },
+      },
+    });
+
+    const isOnlyUserArray = [isOnlyUser];
+
+    if (isOnlyUserArray.length == 1) return RETURN_DATA.LAST_USER;
+
     try {
       await prisma.schoolPersons.delete({
         where: {
@@ -600,8 +638,10 @@ export class SchoolAdminService {
     return RETURN_DATA.SUCCESS;
   }
 
-  async addJoinCode(body: AddJoinCode): Promise<ReturnMessage> {
-    const { schoolUUID, expireDate, name = '', personUUID } = body;
+  async addJoinCode(body: AddJoinCode, token: string): Promise<ReturnMessage> {
+    const { schoolUUID, expireDate, joinCodeName = '' } = body;
+    const jwt = await this.authService.decodeJWT(token);
+    const personUUID = jwt.personUUID;
 
     if (
       !validator.isUUID(schoolUUID.slice(1), 4) ||
@@ -618,7 +658,7 @@ export class SchoolAdminService {
     const nameIsNotAvailable = await prisma.schoolJoinCodes.findFirst({
       where: {
         schoolId: Number(schoolId),
-        joinCodeName: name,
+        joinCodeName: joinCodeName,
       },
     });
 
@@ -645,7 +685,7 @@ export class SchoolAdminService {
               schoolId: Number(schoolId),
             },
           },
-          joinCodeName: name,
+          joinCodeName: joinCodeName,
           joinCode,
           expireDate: new Date(expireDate),
           persons: {
